Web Actions Tab 0.1.3 Async Client Computation
* Move client side processing of metadata to Redux so it is not render blocking
